[RENAME] 테스트코드 없는 컴포넌트는 상위폴더로 이동 #7
